[MOV] gauge: move gauge-specific panel tests
This commit moves the gauge-specific panel tests to a separate file
`gauge_panel_component.test.ts`.

A lot of gauge chart were also kinda duplicated: a test would test that
the input was marked as invalid when entering a wrong value, and another
test would test would test what the error message was when entering an
invalid value. Both can be done in the same test.

Task: 4236214
