+++ tests/figures/chart/charts_component.test.ts
@@ -1268,98 +1268,6 @@ describe("charts", () => {
       }
     );
 
-    test("Can change gauge inflection operator", async () => {
-      createTestChart("gauge");
-      await mountChartSidePanel();
-      await openChartDesignSidePanel(model, env, fixture, chartId);
-
-      expect(model.getters.getChartDefinition(chartId)).toMatchObject({
-        sectionRule: {
-          lowerInflectionPoint: { operator: "<=" },
-          upperInflectionPoint: { operator: "<=" },
-        },
-      });
-
-      const inputs = fixture.querySelectorAll(".o-input[name=operatorType]");
-
-      await setInputValueAndTrigger(inputs[0], "<");
-      expect(model.getters.getChartDefinition(chartId)).toMatchObject({
-        sectionRule: {
-          lowerInflectionPoint: { operator: "<" },
-          upperInflectionPoint: { operator: "<=" },
-        },
-      });
-
-      setInputValueAndTrigger(inputs[1], "<");
-      expect(model.getters.getChartDefinition(chartId)).toMatchObject({
-        sectionRule: {
-          lowerInflectionPoint: { operator: "<" },
-          upperInflectionPoint: { operator: "<" },
-        },
-      });
-    });
-
-    describe("update chart with invalid section rule", () => {
-      beforeEach(async () => {
-        createTestChart("gauge");
-        await mountChartSidePanel();
-        await openChartDesignSidePanel(model, env, fixture, chartId);
-      });
-
-      test("empty rangeMin", async () => {
-        await setInputValueAndTrigger(".o-data-range-min", "");
-        expect(errorMessages()[0]).toEqual(
-          ChartTerms.Errors[CommandResult.EmptyGaugeRangeMin].toString()
-        );
-      });
-
-      test("NaN rangeMin", async () => {
-        await setInputValueAndTrigger(".o-data-range-min", "I'm not a number");
-        expect(errorMessages()[0]).toEqual(
-          ChartTerms.Errors[CommandResult.GaugeRangeMinNaN].toString()
-        );
-      });
-
-      test("empty rangeMax", async () => {
-        await setInputValueAndTrigger(".o-data-range-max", "");
-        expect(errorMessages()[0]).toEqual(
-          ChartTerms.Errors[CommandResult.EmptyGaugeRangeMax].toString()
-        );
-      });
-
-      test("NaN rangeMax", async () => {
-        await setInputValueAndTrigger(".o-data-range-max", "I'm not a number");
-        expect(errorMessages()[0]).toEqual(
-          ChartTerms.Errors[CommandResult.GaugeRangeMaxNaN].toString()
-        );
-      });
-
-      test("rangeMin > rangeMax", async () => {
-        setInputValueAndTrigger(".o-data-range-min", "100");
-
-        await setInputValueAndTrigger(".o-data-range-max", "0");
-        expect(errorMessages()[0]).toEqual(
-          ChartTerms.Errors[CommandResult.GaugeRangeMinBiggerThanRangeMax].toString()
-        );
-      });
-
-      test("NaN LowerInflectionPoint", async () => {
-        await simulateClick(".o-input-lowerInflectionPoint");
-        await setInputValueAndTrigger(".o-input-lowerInflectionPoint", "I'm not a number");
-        expect(errorMessages()[0]).toEqual(
-          ChartTerms.Errors[CommandResult.GaugeLowerInflectionPointNaN].toString()
-        );
-      });
-
-      test("NaN UpperInflectionPoint", async () => {
-        await simulateClick(".o-input-upperInflectionPoint");
-        await setInputValueAndTrigger(".o-input-upperInflectionPoint", "I'm not a number");
-        expect(errorMessages()[0]).toEqual(
-          ChartTerms.Errors[CommandResult.GaugeUpperInflectionPointNaN].toString()
-        );
-      });
-    });
-
     test("Scorecard > error displayed on input fields", async () => {
       createTestChart("scorecard");
       await mountChartSidePanel();
@@ -1377,69 +1285,6 @@ describe("charts", () => {
       await simulateClick(".o-data-labels .o-selection-ok");
       expect(document.querySelector(".o-data-series input")?.classList).toContain("o-invalid");
       expect(document.querySelector(".o-data-labels input")?.classList).toContain("o-invalid");
-    });
-
-    describe("gauge > error displayed on input fields", () => {
-      beforeEach(async () => {
-        createTestChart("gauge");
-        await mountChartSidePanel();
-      });
-
-      test("empty dataRange", async () => {
-        await simulateClick(".o-data-series input");
-        await setInputValueAndTrigger(".o-data-series input", "");
-        await simulateClick(".o-data-series .o-selection-ok");
-        expect(document.querySelector(".o-data-series input")?.classList).toContain("o-invalid");
-      });
-
-      test("empty rangeMin", async () => {
-        await simulateClick(".o-panel-design");
-        await setInputValueAndTrigger(".o-data-range-min", "");
-        expect(document.querySelector(".o-data-range-min")?.classList).toContain("o-invalid");
-      });
-
-      test("NaN rangeMin", async () => {
-        await simulateClick(".o-panel-design");
-        await setInputValueAndTrigger(".o-data-range-min", "bla bla bla");
-        expect(document.querySelector(".o-data-range-min")?.classList).toContain("o-invalid");
-      });
-
-      test("empty rangeMax", async () => {
-        await simulateClick(".o-panel-design");
-        await setInputValueAndTrigger(".o-data-range-max", "");
-        expect(document.querySelector(".o-data-range-max")?.classList).toContain("o-invalid");
-      });
-
-      test("NaN rangeMax", async () => {
-        await simulateClick(".o-panel-design");
-        await setInputValueAndTrigger(".o-data-range-max", "bla bla bla");
-        expect(document.querySelector(".o-data-range-max")?.classList).toContain("o-invalid");
-      });
-
-      test("rangeMin > rangeMax", async () => {
-        await simulateClick(".o-panel-design");
-        setInputValueAndTrigger(".o-data-range-min", "100");
-
-        await setInputValueAndTrigger(".o-data-range-max", "0");
-        expect(document.querySelector(".o-data-range-min")?.classList).toContain("o-invalid");
-        expect(document.querySelector(".o-data-range-max")?.classList).toContain("o-invalid");
-      });
-
-      test("NaN LowerInflectionPoint", async () => {
-        await simulateClick(".o-panel-design");
-        await setInputValueAndTrigger(".o-input-lowerInflectionPoint", "bla bla bla");
-        expect(document.querySelector(".o-input-lowerInflectionPoint")?.classList).toContain(
-          "o-invalid"
-        );
-      });
-
-      test("NaN UpperInflectionPoint", async () => {
-        await simulateClick(".o-panel-design");
-        await setInputValueAndTrigger(".o-input-upperInflectionPoint", "bla bla bla");
-        expect(document.querySelector(".o-input-upperInflectionPoint")?.classList).toContain(
-          "o-invalid"
-        );
-      });
     });
   });
